add `--pass-through` to build and repl (#35)
* add `--pass-through` to build and repl

It should be possible to pass through arguments to the called `purs` commands,
like in this case `compile` and `repl`.
This is needed for e.g. using `purs compile --json-errors` to integrate
psc-package into editors.

* switch from `--pass-through` to implicit passthrough

People think this is a more usable solution. I disagree, but it’s not really
important enough to argue over.