fix(windows): Find purs.exe on Windows, if available
Since `purs` isn't executable on Windows, Turtle procs weren't finding
it to execute.

`purs.cmd` is executable on windows, but it must be run in a shell for
the `~%dp0` macros to expand properly. But running in a shell seems to
break on other platforms (I tested on OS X).

`purs.exe` seems to work just fine if it is on the path. If it is not
already on the path, we have to find it. In this solution, I'm just
looking for the exe relative to where the `purs.cmd` file is and
executing it.