fix(windows): Add compatibility with "purs.cmd"
Windows users who install purescript through npm can't use psc-package
because the `purs` script isn't executable on windows.

To make this work, we need to find the `purs.cmd` that npm installs on
the path and run it in a shell (rather than a process).

This change favors `purs` and `purs.exe` (if available on the path) over
using the cmd.

This also adds a more helpful error message if we can't find a purs
executable.